test: :rotating_light: Add e2e tests for multi-APIs
