+++ .github/workflows/e2e-deploy.yml
@@ -103,6 +103,76 @@ jobs:
         env:
           DATASET: fashion-mnist-784-euclidean.hdf5
           POD_NAME: ${{ steps.deploy_vald.outputs.POD_NAME }}
+  e2e-deploy-lb-multiapis:
+    name: "E2E deploy test (only LB Gateway, multi-APIs)"
+    runs-on: ubuntu-latest
+    timeout-minutes: 60
+    if: startsWith( github.ref, 'refs/tags/') || github.event.action == 'labeled' && github.event.label.name == 'actions/e2e-deploy'
+    steps:
+      - uses: actions/checkout@v2
+      - name: Wait for Docker images
+        if: startsWith( github.ref, 'refs/tags/')
+        id: wait_for_docker_images
+        uses: ./.github/actions/wait-for-docker-image
+      - name: Get PR number
+        id: get_pr_number
+        if: github.event.action == 'labeled' && github.event.label.name == 'actions/e2e-deploy'
+        run: |
+          pr_num=`cat $GITHUB_EVENT_PATH | jq -r ".number"`
+          echo "::set-output name=PR_NUM::${pr_num}"
+      - name: Specify container versions
+        if: github.event.action == 'labeled' && github.event.label.name == 'actions/e2e-deploy'
+        id: specify_container_versions
+        uses: ./.github/actions/detect-docker-image-tags
+        with:
+          tag_name: pr-${{ steps.get_pr_number.outputs.PR_NUM }}
+      - name: Fetch Helm version
+        run: |
+          HELM_VERSION=`make version/helm`
+          echo "::set-output name=helm::${HELM_VERSION}"
+        id: version
+      - uses: rinx/setup-k3d@v0.0.2
+        with:
+          version: latest
+          name: vald
+          agents: 3
+      - name: check k3d
+        run: |
+          kubectl cluster-info
+      - uses: azure/setup-helm@v1
+        with:
+          version: ${{ steps.version.outputs.helm }}
+      - name: Helm version
+        run: |
+          helm version
+      - name: deploy Vald
+        id: deploy_vald
+        uses: ./.github/actions/e2e-deploy-vald
+        with:
+          require_libhdf5: "true"
+          helm_extra_options: ${{ steps.specify_container_versions.outputs.HELM_EXTRA_OPTIONS }}
+          values: .github/helm/values/values-lb.yaml
+          wait_for_selector: app=vald-lb-gateway
+      - name: run E2E CRUD
+        run: |
+          make hack/benchmark/assets/dataset/${DATASET}
+          go test \
+            -v tests/e2e/multiapis/multiapis_test.go \
+            -tags "e2e" \
+            -timeout 30m \
+            -run "TestE2EMultiAPIs" \
+            -host=localhost \
+            -port=8081 \
+            -dataset=`pwd`/hack/benchmark/assets/dataset/${DATASET} \
+            -wait-after-insert=2m \
+            -portforward \
+            -portforward-pod-name=${POD_NAME} \
+            -portforward-pod-port=8081 \
+            -kubeconfig=${KUBECONFIG} \
+            -namespace=default
+        env:
+          DATASET: fashion-mnist-784-euclidean.hdf5
+          POD_NAME: ${{ steps.deploy_vald.outputs.POD_NAME }}
   e2e-deploy-sidecar:
     name: "E2E deploy test (Agent & Sidecar)"
     runs-on: ubuntu-latest
@@ -186,6 +256,7 @@ jobs:
     name: "Slack notification"
     needs:
       - e2e-deploy-lb
+      - e2e-deploy-lb-multiapis
       - e2e-deploy-sidecar
     runs-on: ubuntu-latest
     if: startsWith( github.ref, 'refs/tags/')
